chore(web): splits path-stats object into simpler + full; full tree-shakes out
