+++ common/web/gesture-recognizer/src/engine/index.ts
@@ -2,6 +2,7 @@ import { validateModelDefs } from './headless/gestures/specs/modelDefValidator.j
 
 export { ConstructingSegment } from './headless/subsegmentation/constructingSegment.js';
 export { CumulativePathStats } from './headless/cumulativePathStats.js';
+export { RegressiblePathStats } from './headless/regressiblePathStats.js';
 export { GestureModelDefs } from './headless/gestures/specs/gestureModelDefs.js';
 export { GestureRecognizer } from "./gestureRecognizer.js";
 export { GestureRecognizerConfiguration } from "./configuration/gestureRecognizerConfiguration.js";
